One server verticle to rule them all
